requirements.txt
# ═══════════════════════════════════════════════════════════════
# SenseForge — Multimodal ISAC System Dependencies
# ═══════════════════════════════════════════════════════════════
# NOTE: pyAerial (NVIDIA AI Aerial SDK) must be installed
#   from the SDK repo, NOT from PyPI.
#   See aerial_setup.sh for installation instructions.
# ═══════════════════════════════════════════════════════════════

# ── NVIDIA AI Aerial / Sionna (GPU-accelerated 5G PHY) ───────
sionna>=0.18.0
tensorflow[and-cuda]>=2.13.0

# ── Computer Vision ──────────────────────────────────────────
opencv-python>=4.8.0
ultralytics>=8.0.0
timm>=0.9.0

# ── Deep Learning ────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Scientific Computing ────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
filterpy>=1.4.5

# ── Backend (FastAPI) ───────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
httpx>=0.25.0
requests>=2.31.0
websocket-client>=1.7.0
pillow>=10.0.0

# ── Testing ─────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
